octeontx2-af: fix operand size in bitwise operation
Made size of operands same in bitwise operations.

The patch fixes the klocwork issue, operands in a bitwise operation have
different size at line 375 and 483.

Link: https://lore.kernel.org/r/f4fba33fe4f89b420b4da11d51255e7cc6ea1dbf.1656586269.git.sthotton@marvell.com
(cherry picked from commit b14056914357beee6a84f6ff1b9195f4659fab9d)

Conflicts:
	drivers/net/ethernet/marvell/octeontx2/af/rvu_cpt.c
	- UEK7 contains more recent change to CPT_CTX_ILEN definition
	  through an LTS update

Orabug: 36725601

+++ drivers/net/ethernet/marvell/octeontx2/af/rvu_cpt.c
@@ -484,7 +484,7 @@ static int cpt_inline_ipsec_cfg_inbound(struct rvu *rvu, int blkaddr, u8 cptlf,
 	 */
 	if (!is_rvu_otx2(rvu)) {
 		val = (ilog2(NIX_CHAN_CPT_X2P_MASK + 1) << 16);
-		val |= rvu->hw->cpt_chan_base;
+		val |= (u64)rvu->hw->cpt_chan_base;
 
 		rvu_write64(rvu, blkaddr, CPT_AF_X2PX_LINK_CFG(0), val);
 		rvu_write64(rvu, blkaddr, CPT_AF_X2PX_LINK_CFG(1), val);
